fix for signedness warning
git-svn-id: https://efmsvn2.nrel.gov/lk/svn/trunk@206 beec71b5-fff0-e442-8ac2-4f84a11f714e

+++ src/codegen.cpp
@@ -91,7 +91,7 @@ void code_gen::textout( lk_string &assembly, lk_string &bytecode )
 				}
 				else if ( ip.op == PSH )
 				{
-					static const int MAXWIDTH = 24;
+					static const size_t MAXWIDTH = 24;
 					lk_string nnl( m_constData[ip.arg].as_string() );
 					if ( nnl.size() > MAXWIDTH )
 					{
